# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mlndex-backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/creator/chapters": {
      "post": {
        "tags": [
          "Chapter"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": "@{schema=; encoding=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chapters/{id}": {
      "get": {
        "tags": [
          "Chapter"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "name": "targetId",
            "in": "query",
            "schema": "@{type=integer; format=int32}"
          },
          {
            "name": "targetType",
            "in": "query",
            "schema": "@{$ref=#/components/schemas/CommentTargetType}"
          },
          {
            "name": "page",
            "in": "query",
            "schema": "@{type=integer; format=int32; default=1}"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": "@{type=integer; format=int32; default=20}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/comments/{commentId}": {
      "delete": {
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/moderation-queues/{queueId}/decision": {
      "post": {
        "tags": [
          "ModerationQueues"
        ],
        "parameters": [
          {
            "name": "queueId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/moderation-queues/{queueId}/content-decision": {
      "post": {
        "tags": [
          "ModerationQueues"
        ],
        "parameters": [
          {
            "name": "queueId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/moderation-queues/{queueId}/violation-feedback": {
      "post": {
        "tags": [
          "ModerationQueues"
        ],
        "parameters": [
          {
            "name": "queueId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": "@{type=integer; format=int32; default=1}"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": "@{type=integer; format=int32; default=20}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentTargetType": {
        "enum": [
          "SERIES",
          "CHAPTER"
        ],
        "type": "string"
      },
...
    }
  }
}